+++ examples/gym/convert_policy/Cargo.toml
@@ -9,14 +9,18 @@ anyhow = "1.0.38"
 clap = { version = "4.5.8", features = ["derive"] }
 env_logger = "0.8.2"
 numpy = "0.14.1"
-border-policy-no-backend = { version = "0.0.9", path = "../../../border-policy-no-backend", features = [
-    "tch",
-] }
-border-tch-agent = { version = "0.0.9", path = "../../../border-tch-agent" }
-border-core = { version = "0.0.9", path = "../../../border-core", features = ["tch", "candle"] }
+border-policy-no-backend = { version = "0.0.9", path = "../../../border-policy-no-backend" }
+border-tch-agent = { version = "0.0.9", path = "../../../border-tch-agent", optional = true }
+border-candle-agent = { version = "0.0.9", path = "../../../border-candle-agent", optional = true }
+border-core = { version = "0.0.9", path = "../../../border-core" }
 serde = "1.0.194"
-tch = "0.16.0"
+tch = { version = "0.16.0", optional = true }
+candle-core = { version = "=0.8.4", optional = true }
 bincode = "1.3.3"
 
 [dev-dependencies]
 tempdir = "0.3.7"
+
+[features]
+tch = ["border-policy-no-backend/tch", "border-tch-agent", "border-core/tch", "dep:tch"]
+candle = ["border-policy-no-backend/candle", "border-candle-agent", "border-core/candle", "dep:candle-core"]
