+++ examples/gym/pendulum_std/Cargo.toml
@@ -9,13 +9,11 @@ anyhow = "1.0.38"
 clap = { version = "4.5.8", features = ["derive"] }
 env_logger = "0.8.2"
 numpy = "0.14.1"
-border-policy-no-backend = { version = "0.0.9", path = "../../../border-policy-no-backend", features = [
-    "tch",
-] }
+border-policy-no-backend = { version = "0.0.9", path = "../../../border-policy-no-backend" }
 border-core = { version = "0.0.9", path = "../../../border-core" }
 border-py-gym-env = { version = "0.0.9", path = "../../../border-py-gym-env" }
 serde = "1.0.194"
-tch = "0.16.0"
+# tch = "0.16.0"
 bincode = "1.3.3"
 pyo3 = { version = "=0.14.5", default-features = false }
 ndarray = "0.15.1"
