+++ examples/gym/sac_pendulum/src/main.rs
@@ -49,7 +49,7 @@ const EVAL_INTERVAL: usize = 2_000;
 const REPLAY_BUFFER_CAPACITY: usize = 100_000;
 const N_EPISODES_PER_EVAL: usize = 5;
 const ENV_NAME: &str = "Pendulum-v1";
-const MODEL_DIR: &str = "./model/candle/sac_pendulum";
+const MODEL_DIR: &str = "./model";
 const MLFLOW_EXPERIMENT_NAME: &str = "Gym";
 const MLFLOW_RUN_NAME: &str = "sac-gym-pendulum-v1-candle";
 const MLFLOW_TAGS: &[(&str, &str)] = &[("env", "pendulum"), ("algo", "sac"), ("backend", "candle")];
